rename constant/index file and move i to location folder
